local-dep/services/redirect-azure/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 
azure-functions
pyodbc
redis
pika
